Implementing system calls

+++ libc/dup2.c
@@ -3,5 +3,5 @@
 #include<syscall.h>
 int dup2(int oldfd, int newfd)
 {
-    return syscall_2(SYS_dup2,(uint64_t)oldfd,(uint64_t)newfd);;
+    return syscall_2((uint64_t)SYS_dup2,(uint64_t)oldfd,(uint64_t)newfd);;
 }
